# RunFlow/openapi-mobile-v1.yaml
openapi: 3.0.3
info:
  title: RunFlow Mobile API
  description: |
    RESTful API for the RunFlow Android mobile application.

    ## Authentication
    All endpoints require authentication via JWT token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    The API implements rate limiting per endpoint. Response includes retry headers:
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  version: 1.0.0
  contact:
    name: RunFlow API
servers:
  - url: https://api.runflow.app/api/mobile/v1
    description: Production
  - url: http://localhost:3000/api/mobile/v1
    description: Development
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with Strava
      description: Exchange Strava authorization code for JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /activities/{id}:
    get:
      tags: [Activities]
      summary: Get activity details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Activity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - INVALID_TOKEN
            - TOKEN_EXPIRED
            - FORBIDDEN
            - BAD_REQUEST
            - VALIDATION_ERROR
            - NOT_FOUND
            - CONFLICT
            - RATE_LIMITED
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - DATABASE_ERROR
          description: Machine-readable error code
        details:
          type: object
          additionalProperties: true
          description: Additional error context
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: Request path for debugging
    LoginRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          description: Strava authorization code
    LoginResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - tokenType
        - user
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        tokenType:
          type: string
          enum: [Bearer]
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
          nullable: true
        name:
          type: string
          nullable: true
        image:
          type: string
          format: uri
          nullable: true
        sex:
          type: string
          enum: [MALE, FEMALE, OTHER]
          nullable: true
        birthDate:
          type: string
          format: date
          nullable: true
        hrMax:
          type: integer
          nullable: true
        hrRest:
          type: integer
          nullable: true
        weight:
          type: number
          nullable: true
        height:
          type: number
          nullable: true
        vdotCorrectionFactor:
          type: number
          nullable: true
        lastSyncAt:
          type: string
          format: date-time
          nullable: true
...
    Activity:
      type: object
      properties:
        id:
          type: string
        stravaId:
          type: string
        type:
          type: string
          enum: [RUN, RIDE, VIRTUAL_RIDE, WALK, HIKE, SWIM, WORKOUT, OTHER]
        name:
          type: string
        startDate:
          type: string
          format: date-time
        distance:
          type: number
          description: Distance in meters
        movingTime:
          type: integer
          description: Duration in seconds
        averageSpeed:
          type: number
          nullable: true
        averageHr:
          type: number
          nullable: true
        maxHr:
          type: integer
          nullable: true
        averageCadence:
          type: number
          nullable: true
        hasHeartrate:
          type: boolean
        totalElevation:
          type: number
          description: Elevation gain in meters
        trimp:
          type: number
          nullable: true
        runningTss:
          type: number
          nullable: true
        estimatedVdot:
          type: number
          nullable: true
        trainingType:
          type: string
          nullable: true
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Bad request"
            code: "BAD_REQUEST"
            timestamp: "2024-01-01T00:00:00Z"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Unauthorized"
            code: "UNAUTHORIZED"
            timestamp: "2024-01-01T00:00:00Z"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
            timestamp: "2024-01-01T00:00:00Z"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            timestamp: "2024-01-01T00:00:00Z"